Re-license under GPL verion 2 and Add copyright notice
If you have a problem with this change: You don't have to use this app and you don't have to contribute to it.
But can you do this this easily? As translatable strings are not part of the source code and there have yet to be any contributions to the source code by other people, I am currently the only copyright holder and thus can re-license the project at any point in time without asking you for permission.